Refactor unit and integration tests, remove doc args and returns from all tests
