+++ tests/unit/test_unit_mesan.py
@@ -10,6 +10,7 @@ from pydantic import BaseModel, ConfigDict
 from smhi.constants import MESAN_LEVELS_UNIT, MESAN_PARAMETER_DESCRIPTIONS
 from smhi.mesan import Mesan
 from smhi.models.mesan_model import MesanGeometry, MesanParameterItem, MesanParameters
+from utils import get_response
 
 BASE_URL = (
     "https://opendata-download-metanalys.smhi.se/" + "api/category/mesan2g/version/1/"
@@ -39,13 +40,6 @@ MOCK_MESAN_PARAMETERS = MesanParameters(
 )
 
 
-class MockResponse:
-    def __init__(self, status, header, content):
-        self.status_code = status
-        self.headers = header
-        self.content = content
-
-
 class MockMesanPointData(BaseModel):
     model_config = ConfigDict(arbitrary_types_allowed=True)
 
@@ -59,47 +53,6 @@ class MockMesanMultiPointData(BaseModel):
     approved_time: str
     reference_time: str
     valid_time: str
-
-
-def get_response(file, encode=False):
-    """Read in response.
-
-    Args:
-        file: file to load
-
-    Returns:
-        mocked response
-    """
-    with open(file) as f:
-        mocked_response = f.read()
-
-    headers, content = mocked_response.split("\n\n")
-    status = 200
-    headers = {x.split(":")[0]: x.split(":")[1] for x in headers.split("\n")[1:]}
-
-    if encode is True:
-        content = content.encode("utf-8")
-
-    mocked_get = MockResponse(status, headers, content)
-
-    return mocked_get
-
-
-def get_data(file, load_type=None):
-    """Read in expected data structure.
-
-    Args:
-        file: file to load
-        model
-
-    Returns:
-        expected pydantic model
-    """
-    with open(file) as f:
-        if load_type == "data":
-            return f.read().encode("utf-8")
-
-        return json.load(f)
 
 
 @pytest.fixture
